Address Copilot review: move LoginPage to pages and fix auth.md reference
- Move LoginPage component from src/routes/login.tsx to
  src/components/pages/LoginPage.tsx to comply with architecture rule
  (routes import pages only, logic lives in pages)
- Slim src/routes/login.tsx down to route definition + beforeLoad only
- Fix docs/specs/features/auth.md to reference the existing LoginPage
  instead of the non-existent AuthForm organism
